third_party/blink/web_tests/external/wpt/css/css-values/sin-cos-tan-computed-expected.txt

This is a testharness.js-based test.
FAIL cos(0) should be used-value-equivalent to 1 assert_equals: cos(0) and 1 serialize to the same thing in used values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL sin(0) should be used-value-equivalent to 0 assert_equals: sin(0) and 0 serialize to the same thing in used values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
FAIL tan(0) should be used-value-equivalent to 0 assert_equals: tan(0) and 0 serialize to the same thing in used values. expected "matrix(0, 0, 0, 0, 0, 0)" but got "none"
FAIL calc(sin(pi/2 - pi/2) ) should be used-value-equivalent to 0 Cannot read properties of undefined (reading 'split')
FAIL calc(cos(pi - 3.14159265358979323846) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(cos(e - 2.7182818284590452354) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(sin(30deg + 1.0471967rad ) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
PASS calc(cos(30deg - 0.523599rad ) ) should be used-value-equivalent to 1
FAIL calc(sin(3.14159 / 2 + 1 - 1) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(sin(100grad) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(cos(0 / 2 + 1 - 1) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(tan(30deg + 0.261799rad ) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(tan(0.7853975rad ) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(tan(3.14159 / 4 + 1 - 1) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(sin(0.25turn) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(cos(sin(cos(pi) + 1))) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
FAIL calc(sin(tan(pi/4)*pi/2) ) should be used-value-equivalent to 1 Cannot read properties of undefined (reading 'split')
Harness: the test ran to completion.
